# ENH: Optimize np.power(x, 2) for double and float type