--- a/numpy/_core/src/umath/loops_umath_fp.dispatch.c.src
+++ b/numpy/_core/src/umath/loops_umath_fp.dispatch.c.src
@@ -271,16 +271,6 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
         return;
     }
 #endif
-    if (stride_zero) {
-        // this maybe a bit faster than the generic loop, but not sure
-        BINARY_DEFS_ZERO_STRIDE
-        const @type@ in2 = *(@type@ *)ip2;
-        BINARY_LOOP_SLIDING_ZERO_STRIDE {
-            const @type@ in1 = *(@type@ *)ip1;
-            *(@type@ *)op1 = npy_@intrin@@vsub@(in1, in2);
-        }
-        return;
-    }
     BINARY_LOOP {
         const @type@ in1 = *(@type@ *)ip1;
         const @type@ in2 = *(@type@ *)ip2;