--- a/numpy/_core/src/umath/loops_umath_fp.dispatch.c.src
+++ b/numpy/_core/src/umath/loops_umath_fp.dispatch.c.src
@@ -237,10 +237,10 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@func@)
 {
     int stride_zero = steps[1]==0;
     if (stride_zero) {
-        BINARY_DEFS_ZERO_STRIDE
+        BINARY_DEFS
         const @type@ in2 = *(@type@ *)ip2;
         if (in2 == 2.0) {
-            BINARY_LOOP_SLIDING_ZERO_STRIDE {
+            BINARY_LOOP_SLIDING {
                 const @type@ in1 = *(@type@ *)ip1;
                 *(@type@ *)op1 = in1 * in1;
             }