--- a/numpy/_core/tests/test_umath.py
+++ b/numpy/_core/tests/test_umath.py
@@ -1235,6 +1235,16 @@ class TestPower:
             r = np.array([1, 1, np.inf, 0, np.inf, 0, np.inf, 0], dt)
             assert_equal(np.power(a, b), r)
 
+    def test_power_fast_paths(self):
+        # gh-26055
+        for dt in [np.float32, np.float64]:
+            expected = np.array([0.0, 1.21, 4., 1.44e+26, np.inf, np.inf])
+            a = np.array([0, 1.1, 2, 12e12, np.inf, -np.inf], dt)
+            assert_equal(np.power(a, 2.), expected.astype(dt))
+
+            expected = np.sqrt(a)
+            assert_equal(np.power(a, 0.5), expected.astype(dt))
+
 
 class TestFloat_power:
     def test_type_conversion(self):