We optimize `np.power(x, e)` by adding a fast path for the case `e=2` and `x` is double or float.

Updated benchmark (linux):

Main
```
x**1 54.387 [us]
x**2 50.409 [us]
x**3 1999.226 [us]
x**1.123 1982.337 [us]
x**0.5 206.666 [us]
power(x, 1) 1980.208 [us]
power(x, 2) 1987.717 [us]
power(x, 3) 1982.347 [us]
power(x, 1.123) 1979.270 [us]
power(x, 0.5) 1981.093 [us]
power(x_float32, 1) 917.876 [us]
power(x_float32, 2) 937.346 [us]
power(x_float32, 3) 1006.258 [us]
power(x_float32, 1.123) 916.753 [us]
power(x_float32, 0.5) 918.313 [us]
power(double_size10, 1) 1.116 [us]
power(double_size10, 2) 1.127 [us]
power(double_size10, 3) 1.172 [us]
power(double_size10, 1.123) 1.010 [us]
power(double_size10, 0.5) 0.941 [us]
```
PR
```
x**1 55.415 [us]
x**2 51.346 [us]
x**3 1953.280 [us]
x**1.123 1953.009 [us]
x**0.5 206.721 [us]
power(x, 1) 1947.996 [us]
power(x, 2) 58.777 [us]
power(x, 3) 1951.243 [us]
power(x, 1.123) 1949.528 [us]
power(x, 0.5) 406.441 [us]
power(x_float32, 1) 900.470 [us]
power(x_float32, 2) 53.788 [us]
power(x_float32, 3) 900.063 [us]
power(x_float32, 1.123) 903.382 [us]
power(x_float32, 0.5) 205.104 [us]
power(double_size10, 1) 1.087 [us]
power(double_size10, 2) 0.912 [us]
power(double_size10, 3) 1.218 [us]
power(double_size10, 1.123) 0.918 [us]
power(double_size10, 0.5) 0.759 [us]
```
<details><summary>Benchmark script</summary>

```
import sys
del(sys.path[0])
import numpy as np
print(np, np.__version__)
np.show_config()

import timeit

exponents = [1, 2, 3, 1.123, .5]

xf=np.arange(100_000.).astype(np.float64)
nf = 1_000
for e in exponents:
    dt=timeit.timeit(f'xf**{e}', globals=globals(), number=nf)
    print(f'x**{e} {1e6*dt/nf:.3f} [us]')

for e in exponents:
    dt=timeit.timeit(f'np.power(xf, {e})', globals=globals(), number=nf)
    print(f'power(x, {e}) {1e6*dt/nf:.3f} [us]')

f=np.arange(100_000.).astype(np.float32)
for e in exponents:
    dt=timeit.timeit(f'np.power(f, {e})', globals=globals(), number=nf)
    print(f'power(x_float32, {e}) {1e6*dt/nf:.3f} [us]')


nf = 1_000_000
y=np.arange(10.)
for e in exponents:
    dt=timeit.timeit(f'np.power(y, {e})', globals=globals(), number=nf)
    print(f'power(double_size10, {e}) {1e6*dt/nf:.3f} [us]')
```
</details>

Old benchmark (windows)
```
import numpy as np
import timeit

xf=np.arange(100_000+.0)
nf = 2_000

for e in [1, 2, 3]:
    dt=timeit.timeit(f'xf**{e}', globals=globals(), number=nf)
    print(f'x**{e} {1e6*dt/nf:.3f} [us]')

for e in [1, 2, 3]:
    dt=timeit.timeit(f'np.power(xf, {e})', globals=globals(), number=nf)
    print(f'power(x, {e}) {1e6*dt/nf:.3f} [us]')
dt=timeit.timeit('np.power(xf, 2.)', globals=globals(), number=nf)
print(f'power(x, 2.) {1e6*dt/nf:.3f} [us]')
```
Results on main:
```
x**1 131.168 [us]
x**2 75.814 [us]
x**3 1509.223 [us]
power(x, 1) 306.960 [us]
power(x, 2) 1620.303 [us]
power(x, 3) 1571.338 [us]
power(x, 2.) 1493.646 [us]
```
Results on this PR:
```
x**1 119.589 [us]
x**2 77.959 [us]
x**3 1445.394 [us]
power(x, 1) 257.166 [us]
power(x, 2) 58.722 [us]
power(x, 3) 1406.567 [us]
power(x, 2.) 57.718 [us]
```

Notes:

* Maybe the fast path for exponent 2 should be outside the SIMD check. On my system SIMD is not selected/available, so I cannot benchmark this.
* Benchmark results are a bit unstable on my machine, so it would be good to reproduce this
* We could special case also other often used exponents. For example -1, 0, .5, 1, 3, 4. But we should check whether that gives the exact same results as the current implementation.
* Depending on the compiler the option `ffast-math` can make `pow` much faster for special cases such as `e=2`. See for example https://stackoverflow.com/a/2940800

Related: #26045, #9363, #6399, #22651, #7786

<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->