@eendebakpt 2024-03-16
All tests pass, but this PR does change behavior. A minimal example:
```
z=np.frombuffer(b'\xb6\xfe\x10\x16\xe1\xa7\xda?')
print(z, np.power(z, 2) - z*z)
```
On main `np.power(z, 2)` and `z*z` slightly differ, with this PR they are exactly equal.