@ngoldbaum 2024-03-18
Given the dramatic performance improvement, I think accepting some small change in output is ok. Is it possible to hint to the compiler to generalize this a bit? Or maybe manually unroll up to e.g. `n=5`?

It would be good to see some benchmarks to capture the overhead of the extra branching for non-integer powers, if any.

Added the triage review label to make sure this gets looked at during a triage meeting.