@r-devulap 2024-03-18
> We optimize np.power(x, e) by adding a fast path for the case e=2 and x is double.

Your patch seems to do this for float as well. Or am I missing something?