@eendebakpt 2024-03-19
> Given the dramatic performance improvement, I think accepting some small change in output is ok. Is it possible to hint to the compiler to generalize this a bit? Or maybe manually unroll up to e.g. `n=5`?
> 
> It would be good to see some benchmarks to capture the overhead of the extra branching for non-integer powers, if any.

I added some tests with smaller arrays to see whether the overhead impacts the non-integer case (results updated in first comment). It seems the impact is small or none, but it might depend on the compiler/platform and would be good to benchmark on an independent system.

Adding more special cases is possible, but we need to decide which cases are important as each added case adds a bit more complexity (and overhead). I now have `e=2` and `e=.5`. Other cases to consider are -1 and -2 (but this might require some though about whether cases like 0 are handles correctly) and indeed `e=3, 4, 5,  ...`