@seberg 2024-03-19
The impact is largest in some degenerate cases.  Small arrays are dominated by other overheads anyway.
Usage of `where=[0, 1] * N` may be the worst but is rather niche and ridiculous.  Things like `arr = np.empty((large, 2))[::2, :]` might run into it (but I am actually not sure, it may be that it goes into buffered paths making the inner-loop large).