@r-devulap 2024-03-19
> Adding more special cases is possible, but we need to decide which cases are important

@eendebakpt out of curiosity, why optimize for `np.power(x, 2)`? one could as well just use `x * x`,