@eendebakpt 2024-03-19
The following script can be used to generate some cases where `z*z` differs from `np.power(z, e)` (results might be platform dependent).
```
import sympy
import numpy as np
import random
from codecs import decode
import struct


def bin_to_float(b):
    """ Convert binary string to a float. """
    bf = int(b, 2).to_bytes(8)  # 8 bytes needed for IEEE 754 binary64.
    return struct.unpack('>d', bf)[0]

def check_float(z: float, e: int = 3):
    symbolic_z= sympy.sympify(z)

    symbolic_product = 1
    z_product = 1
    for _ in range(e):
        symbolic_product *= symbolic_z
        z_product *= z
    delta = symbolic_product - z_product
    delta_pow = symbolic_product - np.power(z, e).item()
    if  delta_pow!=delta:
       
        print(f'{z}: delta product {delta}, delta pow(., {e}) {delta_pow}')
    return delta_pow!=delta

z=np.frombuffer(b'\xb6\xfe\x10\x16\xe1\xa7\xda?')
z=float(z.item())
check_float(z)

x=np.random.rand(40_000,)
n=0
for z in x:
    symbolic_z= sympy.sympify(z)
    n+= check_float(z)
print(f'{n}/{x.size} = {n/x.size}')

for ii in range(10_000):
    z = bin_to_float(bin(random.getrandbits(64)))
    check_float(z)
```
In the cases generated with this script the values for `z*z` agree well with `np.power(z, 2)`, but already for `e=3` there are large differences. For example for `z=-15656763.486516586`. In the cases generated the calculated product is a better approximation than `np.power(z, 3)`.