@eendebakpt 2024-03-19
> > Adding more special cases is possible, but we need to decide which cases are important
> 
> @eendebakpt out of curiosity, why optimize for `np.power(x, 2)`? one could as well just use `x * x`,

That is a fair point. There are several issues (links in the first comment) where users experience a slow np.power (also for other exponents), and this would be one way to provide a performance improvement. Replacing `np.power(x, 2)` with `x*x` is not always possible or elegant (for example when `np.power` is used inside a method where the exponent is one of the arguments).

But whether numpy should special case for certain exponents (and if so which ones) is not clear to me. It is on the agenda for the next triage meeting I believe, that also seems like a good place to discuss.