@mhvk 2024-03-19
> @eendebakpt out of curiosity, why optimize for `np.power(x, 2)`? one could as well just use `x * x`,

Numpy already optimizes `x ** 2` to call `np.square` - I think in any higher level language one wants to try to avoid people having to think about performance too much, but rather focus on code correctness.