@eendebakpt 2024-03-21
@ngoldbaum Thanks for the feedback! I implemented `e=2` and `e=.5`. 

I tried some experiments to determine the accuracy. Not sure there are representative, but here are some results.

For `e=3` the precision loss seems small (cases found so far are always within one ULP), but the results can be less precise. So unless there is an expert that can confirm adding `e=3` is ok, I will leave it out of this PR. One example for type float32 (the blue data points are a bit further from the exact values than the red ones):
![image](https://github.com/numpy/numpy/assets/883786/385eae97-7eda-4f58-b846-530f58bcb0ed)

For `e=0.5` all the differences between `sqrt` and `np.power` seem to equal far off from the exact value (and all within one ULP). An example: 
![image](https://github.com/numpy/numpy/assets/883786/900efd46-e91d-4eb0-89c6-59c1b5488cd7)