@r-devulap 2024-04-04
Added benchmarks in https://github.com/numpy/numpy/pull/26212. Results for this PR on Intel SKX: 

```
| Change   | Before [45173782] <main>   | After [b5995034] <power>   |   Ratio | Benchmark (Parameter)                                          |
|----------|----------------------------|----------------------------|---------|----------------------------------------------------------------|
| +        | 8.01±0.09ms                | 8.51±0.01ms                |    1.06 | bench_ufunc.BinaryBench.time_pow(<class 'numpy.float32'>)      |
| -        | 8.17±0.01ms                | 928±1μs                    |    0.11 | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float32'>) |
| -        | 19.3±0.4ms                 | 1.90±0ms                   |    0.1  | bench_ufunc.BinaryBench.time_pow_half(<class 'numpy.float64'>) |
| -        | 8.23±0.02ms                | 625±2μs                    |    0.08 | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float32'>)    |
| -        | 19.3±0.4ms                 | 857±20μs                   |    0.04 | bench_ufunc.BinaryBench.time_pow_2(<class 'numpy.float64'>)    |
```