@eendebakpt 2024-04-04
> Minor changes but mostly LGTM. Do you know if there are unit tests that cover these special cases for `np.power`? If not, it might be worth adding them to test this new code paths.
To be explicit I added a test for the two special cases.