@ngoldbaum 2024-04-04
Is the 6% performance drop in `test_pow` statistically significant? What about 64 bit?