@r-devulap 2024-04-05
> Is the 6% performance drop in `test_pow` statistically significant? What about 64 bit?
 
we see this often. when the code placement changes it sometimes affects perf a little bit, But I am not worried about 6% too much. I have tried multiple times to narrow down why that happens but failed every time :/